requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.9.1
cffi==1.17.1
channels==4.3.1
channels_redis==4.3.0
Django==5.2.5
django-allauth==65.11.2
django-cors-headers==4.7.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
epicstuff==0.1.8
markdown-it-py==4.0.0
mdurl==0.1.2
msgpack==1.1.1
pillow==11.3.0
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg2-binary==2.9.10
pycparser==2.22
Pygments==2.19.2
PyJWT==2.10.1
python-dotenv==1.1.1
PyYAML==6.0.2
redis==6.4.0
rich==14.1.0
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.12
sqlparse==0.5.3
taml==1.0.0
typing_extensions==4.15.0
